citibike-pipeline/assets/ingestion/requirements.txt
google-cloud-storage
requests
polars

